1 версия расчета тестирования

+++ SOTA/Models/SotaDB.cs
@@ -116,7 +116,7 @@ namespace SOTA.Models
         public int IdRabota { get; set; }
         public int IdUser { get; set; }
         public int IdZadania { get; set; }
-        public int Ball { get; set; }
+        public double Ball { get; set; }
         public DateTime Date { get; set; }
